apps/market-watcher/requirements.txt
fastapi
uvicorn
google-generativeai
python-dotenv
requests
feedparser
beautifulsoup4
pydantic
